refactor instrument lookup route
